+++ systemimage/tests/test_state.py
@@ -16,8 +16,10 @@
 """Test the state machine."""
 
 __all__ = [
-    'TestState',
+    'TestCommandFileDelta',
+    'TestCommandFileFull',
     'TestRebootingState',
+    'TestState',
     ]
 
 
@@ -304,8 +306,12 @@ class TestState(unittest.TestCase):
         self.assertRaises(SignatureError, next, state)
 
 
-class TestRebooting(unittest.TestCase):
-    """Test various state transitions leading to a reboot."""
+class _StateTestsBase(unittest.TestCase):
+    # Must override in base classes.
+    INDEX_FILE = None
+
+    # For more detailed output.
+    maxDiff = None
 
     def setUp(self):
         self._stack = ExitStack()
@@ -324,9 +330,9 @@ class TestRebooting(unittest.TestCase):
             index_path = os.path.join(
                 self._serverdir, 'stable', 'nexus7', 'index.json')
             head, tail = os.path.split(index_path)
-            copy('index_13.json', head, tail)
+            copy(self.INDEX_FILE, head, tail)
             sign(index_path, 'device-signing.gpg')
-            setup_index('index_13.json', self._serverdir, 'device-signing.gpg')
+            setup_index(self.INDEX_FILE, self._serverdir, 'device-signing.gpg')
         except:
             self._stack.close()
             raise
@@ -357,6 +363,12 @@ class TestRebooting(unittest.TestCase):
             dict(type='device-signing'),
             os.path.join(self._serverdir, 'stable', 'nexus7',
                          'device-signing.tar.xz'))
+
+
+class TestRebooting(_StateTestsBase):
+    """Test various state transitions leading to a reboot."""
+
+    INDEX_FILE = 'index_13.json'
 
     @testable_configuration
     def test_keyrings_copied_to_upgrader_paths(self):
@@ -431,23 +443,6 @@ class TestRebooting(unittest.TestCase):
         self.assertTrue(got_reboot)
 
     @testable_configuration
-    def test_reboot_command_file(self):
-        # The command file gets properly filled.
-        self._setup_keyrings()
-        list(State())
-        path = os.path.join(config.updater.cache_partition, 'ubuntu_command')
-        with open(path, 'r', encoding='utf-8') as fp:
-            command = fp.read()
-        self.assertMultiLineEqual(command, """\
-load_keyring image-master.tar.xz image-master.tar.xz.asc
-load_keyring image-signing.tar.xz image-signing.tar.xz.asc
-load_keyring device-signing.tar.xz device-signing.tar.xz.asc
-update 6.txt 6.txt.asc
-update 7.txt 7.txt.asc
-update 5.txt 5.txt.asc
-""")
-
-    @testable_configuration
     def test_run_until(self):
         # It is possible to run the state machine either until some specific
         # state is completed, or it runs to the end.
@@ -478,3 +473,53 @@ update 5.txt 5.txt.asc
                 'systemimage.tests.reboot.TestableReboot.reboot', reboot_mock):
             list(state)
         self.assertTrue(got_reboot)
+
+
+class TestCommandFileFull(_StateTestsBase):
+    INDEX_FILE = 'index_13.json'
+
+    @testable_configuration
+    def test_full_command_file(self):
+        # A full update's command file gets properly filled.
+        self._setup_keyrings()
+        State().run_until('reboot')
+        path = os.path.join(config.updater.cache_partition, 'ubuntu_command')
+        with open(path, 'r', encoding='utf-8') as fp:
+            command = fp.read()
+        self.assertMultiLineEqual(command, """\
+load_keyring image-master.tar.xz image-master.tar.xz.asc
+load_keyring image-signing.tar.xz image-signing.tar.xz.asc
+load_keyring device-signing.tar.xz device-signing.tar.xz.asc
+format system
+mount system
+update 6.txt 6.txt.asc
+update 7.txt 7.txt.asc
+update 5.txt 5.txt.asc
+unmount system
+""")
+
+
+class TestCommandFileDelta(_StateTestsBase):
+    INDEX_FILE = 'index_15.json'
+
+    @testable_configuration
+    def test_delta_command_file(self):
+        # A delta update's command file gets properly filled.
+        self._setup_keyrings()
+        # Set the current build number so a delta update will work.
+        with open(config.system.build_file, 'w', encoding='utf-8') as fp:
+            print(20120100, file=fp)
+        State().run_until('reboot')
+        path = os.path.join(config.updater.cache_partition, 'ubuntu_command')
+        with open(path, 'r', encoding='utf-8') as fp:
+            command = fp.read()
+        self.assertMultiLineEqual(command, """\
+load_keyring image-master.tar.xz image-master.tar.xz.asc
+load_keyring image-signing.tar.xz image-signing.tar.xz.asc
+load_keyring device-signing.tar.xz device-signing.tar.xz.asc
+mount system
+update 6.txt 6.txt.asc
+update 7.txt 7.txt.asc
+update 5.txt 5.txt.asc
+unmount system
+""")